server/requirements.txt
scrapy>=2.11.0
pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

